requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy
psycopg2-binary
PyMuPDF
langchain
sentence-transformers
chromadb
requests
